requirements.txt
openai>=1.0.0
pypdf
langchain-text-splitters
sentence-transformers
faiss-cpu